+++ noodles-bam/src/lazy/record/bounds.rs
@@ -2,7 +2,7 @@ use std::ops::{Range, RangeFrom};
 
 pub const REFERENCE_SEQUENCE_ID_RANGE: Range<usize> = 0..4;
 pub const ALIGNMENT_START_RANGE: Range<usize> = 4..8;
-pub const READ_NAME_LENGTH_INDEX: usize = 8;
+pub const NAME_LENGTH_INDEX: usize = 8;
 pub const MAPPING_QUALITY_INDEX: usize = 9;
 pub const CIGAR_OP_COUNT_RANGE: Range<usize> = 12..14;
 pub const FLAGS_RANGE: Range<usize> = 14..16;
@@ -13,19 +13,19 @@ pub const TEMPLATE_LENGTH_RANGE: Range<usize> = 28..32;
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct Bounds {
-    pub read_name_end: usize,
+    pub name_end: usize,
     pub cigar_end: usize,
     pub sequence_end: usize,
     pub quality_scores_end: usize,
 }
 
 impl Bounds {
-    pub fn read_name_range(&self) -> Range<usize> {
-        TEMPLATE_LENGTH_RANGE.end..self.read_name_end
+    pub fn name_range(&self) -> Range<usize> {
+        TEMPLATE_LENGTH_RANGE.end..self.name_end
     }
 
     pub fn cigar_range(&self) -> Range<usize> {
-        self.read_name_end..self.cigar_end
+        self.name_end..self.cigar_end
     }
 
     pub fn sequence_range(&self) -> Range<usize> {