+++ noodles-bam/src/lazy/record/name.rs
@@ -2,16 +2,16 @@ use std::io;
 
 use noodles_sam as sam;
 
-/// A raw BAM record read name.
+/// A raw BAM record name.
 #[derive(Debug, Eq, PartialEq)]
-pub struct ReadName<'a>(&'a [u8]);
+pub struct Name<'a>(&'a [u8]);
 
-impl<'a> ReadName<'a> {
+impl<'a> Name<'a> {
     pub(super) fn new(src: &'a [u8]) -> Self {
         Self(src)
     }
 
-    /// Returns the read name as a byte slice.
+    /// Returns the name as a byte slice.
     ///
     /// The returned slice will _not_ have the trailing `NUL` terminator.
     pub fn as_bytes(&self) -> &[u8] {
@@ -20,23 +20,23 @@ impl<'a> ReadName<'a> {
     }
 }
 
-impl<'a> sam::alignment::record::Name for ReadName<'a> {
+impl<'a> sam::alignment::record::Name for Name<'a> {
     fn as_bytes(&self) -> &[u8] {
         self.as_bytes()
     }
 }
 
-impl<'a> AsRef<[u8]> for ReadName<'a> {
+impl<'a> AsRef<[u8]> for Name<'a> {
     fn as_ref(&self) -> &[u8] {
         self.0
     }
 }
 
-impl<'a> TryFrom<ReadName<'a>> for sam::record::Name {
+impl<'a> TryFrom<Name<'a>> for sam::record::Name {
     type Error = io::Error;
 
-    fn try_from(bam_read_name: ReadName<'a>) -> Result<Self, Self::Error> {
-        Self::try_from(bam_read_name.as_bytes().to_vec())
+    fn try_from(bam_name: Name<'a>) -> Result<Self, Self::Error> {
+        Self::try_from(bam_name.as_bytes().to_vec())
             .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     }
 }
@@ -47,10 +47,10 @@ mod tests {
 
     #[test]
     fn test_as_bytes() {
-        let read_name = ReadName::new(b"r0\x00");
-        assert_eq!(read_name.as_bytes(), b"r0");
+        let name = Name::new(b"r0\x00");
+        assert_eq!(name.as_bytes(), b"r0");
 
-        let read_name = ReadName::new(b"r0");
-        assert_eq!(read_name.as_bytes(), b"r0");
+        let name = Name::new(b"r0");
+        assert_eq!(name.as_bytes(), b"r0");
     }
 }